# services/gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Neobank Gateway API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /v1/transfers:
    get:
      operationId: listTransfers
      parameters:
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/XUserId"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTransfer
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/XUserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransferRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
        "200":
          description: Completed (idempotent)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Transfer failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
        "502":
          description: Bad Gateway
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
    Authorization:
      name: Authorization
      in: header
      schema:
        type: string
    XUserId:
      name: X-User-Id
      in: header
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CreateTransferRequest:
      type: object
      required: [recipient_phone, amount]
      properties:
        recipient_phone:
          type: string
        amount:
          type: string
        currency:
          type: string
        memo:
          type: string
    Transfer:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        sender_user_id:
          type: string
        recipient_user_id:
          type: string
        amount:
          type: string
        currency:
          type: string
        ledger_transfer_id:
          type: string
        failure_reason:
          type: string
        memo:
          type: string
    TransferList:
      type: object
      required: [transfers]
      properties:
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/Transfer"
...